Implement the BaseUnoTheme abstract class (#1324)
* chore: move SharedColors/Palette to Uno.Themes

* feat: created BaseUnoTheme and use it on Uno.Material

* chore: rename BaseUnoTheme -> BaseTheme

* chore: adds Uno.Theme as reference and generate markup for it

* chore: changed the markup generation to WinUI.Markup project

* chore: return just one ResourceDictionary from abstract method

* chore: forcing Markup package on Uno.Themes

* fix: Fix Markup generator on Themes.WinUI.Markup
