+++ src/library/Uno.Material.WinUI.Markup/AssemblyInfo.cs
@@ -2,4 +2,5 @@
 using Uno.Extensions.Markup.Generator;
 
 [assembly: GenerateMarkupForAssembly(typeof(Uno.Material.MaterialResources))]
+
 [assembly: AssemblyMetadata("IsTrimmable", "True")]
