+++ src/library/Uno.Material.WinUI.Markup/Uno.Material.WinUI.Markup.csproj
@@ -16,6 +16,7 @@
 	<ItemGroup>
 		<ProjectReference Include="..\Uno.Material\Uno.Material.csproj" />
 		<ProjectReference Include="..\Uno.Themes.WinUI.Markup\Uno.Themes.WinUI.Markup.csproj" />
+		<ProjectReference Include="..\Uno.Themes\Uno.Themes.csproj" />
 	</ItemGroup>
 
 	<ItemGroup>
