+++ src/library/Uno.Themes.WinUI.Markup/Uno.Themes.WinUI.Markup.csproj
@@ -16,6 +16,11 @@
 	<ItemGroup>
 		<PackageReference Include="Uno.WinUI.Markup" Version="5.0.13" />
 		<PackageReference Include="Uno.WinUI" Version="5.0.19" />
+		<PackageReference Include="Uno.Extensions.Markup.Generators" PrivateAssets="All" Version="5.0.13" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\Uno.Themes\Uno.Themes.csproj" />
 	</ItemGroup>
 
 	<PropertyGroup>
