+++ src/library/Uno.Themes/Styles/Applications/Common/SharedColorPalette.xaml
