+++ src/library/Uno.Themes/Styles/Applications/Common/SharedColors.xaml
