+++ src/library/Uno.Themes/Uno.Themes.csproj
@@ -98,6 +98,16 @@
 	</ItemGroup>
 
 	<ItemGroup>
+	  <None Remove="Styles\Applications\Common\SharedColorPalette.xaml" />
+	  <None Remove="Styles\Applications\Common\SharedColors.xaml" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <UpToDateCheckInput Remove="Styles\Applications\Common\SharedColorPalette.xaml" />
+	  <UpToDateCheckInput Remove="Styles\Applications\Common\SharedColors.xaml" />
+	</ItemGroup>
+
+	<ItemGroup>
 		<PackageReference Include="Uno.Core.Extensions.Disposables" Version="4.0.1" />
 		<PackageReference Include="Uno.Core.Extensions.Logging.Singleton" Version="4.0.1" />
 		<PackageReference Include="Uno.Fonts.Roboto" Version="2.2.2" />
